requirements.txt
numpy
tensorflow==2.15.1
streamlit>=1.25.0
pillow
scikit-learn
plotly
pyyaml
pandas